requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
beautifulsoup4>=4.12.0
pydantic>=2.6.0
procyclingstats~=0.2.8
requests~=2.32.5

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
supabase>=2.0.0

# Caching & scheduling
apscheduler>=3.10.0

# NLP Layer 1
rapidfuzz>=3.6.0

# Moderation
better-profanity>=0.7.0

# Calendar (deferred to Plan 2)
# icalendar>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.20.0  # SQLite async driver for tests